Add dir check

+++ my_links.sh
@@ -20,10 +20,13 @@ echo
 CREATE_SYMLINK() {
 	
 	echo
-	
-	rm -rf "${LINK_DIR}/${LINK_FILE}" || true
-	ln -s "${SOURCE_DIR}/${SOURCE_FILE}" "${LINK_DIR}/${LINK_FILE}"
-	
+
+	if [ ! -d "${LINK_DIR}/${LINK_FILE}" ]
+        then
+	    rm -rf "${LINK_DIR}/${LINK_FILE}" || true
+	    ln -s "${SOURCE_DIR}/${SOURCE_FILE}" "${LINK_DIR}/${LINK_FILE}"
+	fi
+
 	echo
 	
 }
@@ -32,9 +35,12 @@ SUDO_CREATE_SYMLINK() {
 	
 	echo
 	
-	sudo rm -rf "${LINK_DIR}/${LINK_FILE}" || true
-	sudo ln -s "${SOURCE_DIR}/${SOURCE_FILE}" "${LINK_DIR}/${LINK_FILE}"
-	
+        if [ ! -d "${LINK_DIR}/${LINK_FILE}" ]
+        then
+	    sudo rm -rf "${LINK_DIR}/${LINK_FILE}" || true
+	    sudo ln -s "${SOURCE_DIR}/${SOURCE_FILE}" "${LINK_DIR}/${LINK_FILE}"
+	fi
+
 	echo
 	
 }
@@ -43,9 +49,12 @@ COPY_FILE() {
 	
 	echo
 	
-	rm -rf "${LINK_DIR}/${LINK_FILE}" || true
-	cp "${SOURCE_DIR}/${SOURCE_FILE}" "${LINK_DIR}/${LINK_FILE}"
-	
+        if [ ! -d "${LINK_DIR}/${LINK_FILE}" ]
+        then
+	    rm -rf "${LINK_DIR}/${LINK_FILE}" || true
+	    cp "${SOURCE_DIR}/${SOURCE_FILE}" "${LINK_DIR}/${LINK_FILE}"
+	fi
+
 	echo
 	
 }
@@ -54,9 +63,12 @@ SUDO_COPY_FILE() {
 	
 	echo
 	
-	sudo rm -rf "${LINK_DIR}/${LINK_FILE}" || true
-	sudo cp "${SOURCE_DIR}/${SOURCE_FILE}" "${LINK_DIR}/${LINK_FILE}"
-	
+        if [ ! -d "${LINK_DIR}/${LINK_FILE}" ]
+        then
+	    sudo rm -rf "${LINK_DIR}/${LINK_FILE}" || true
+	    sudo cp "${SOURCE_DIR}/${SOURCE_FILE}" "${LINK_DIR}/${LINK_FILE}"
+	fi
+
 	echo
 	
 }
